use-cases/03-conversational-agent/python/pyproject.toml
[project]
name = "conversational-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.10",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]
